domain model structure refactor
