+++ data/src/main/java/com/example/data/backend/AppRepositoryImpl.kt
@@ -34,37 +34,37 @@ import com.example.data.backend.models.SendResetCodeRequest
 import com.example.data.backend.models.Tokens
 import com.example.data.tokenmanager.TokenManager
 import com.example.data.verifycodemanager.VerifyCodeManager
-import com.example.domain.entity.DataCompleteResponseEntity
-import com.example.domain.entity.DataEmailResetEntity
-import com.example.domain.entity.DataSendCodeDomain
-import com.example.domain.entity.EmailPassDataErrorEntity
-import com.example.domain.entity.EmailPassResetErrorEntity
-import com.example.domain.entity.EmailPassResetErrorsEntity
-import com.example.domain.entity.EmailResetResponseEntity
-import com.example.domain.entity.EmailResetResultEntity
-import com.example.domain.entity.ErrorResponse
-import com.example.domain.entity.LoginData
-import com.example.domain.entity.LoginDataError
-import com.example.domain.entity.LoginErrorsDomain
-import com.example.domain.entity.LoginResponse
-import com.example.domain.entity.LoginResultEntity
-import com.example.domain.entity.LoginTokens
-import com.example.domain.entity.RegistrationDataEntity
-import com.example.domain.entity.RegistrationErrorDetailEntity
-import com.example.domain.entity.RegistrationErrorEntity
-import com.example.domain.entity.RegistrationErrorsDataEntity
-import com.example.domain.entity.RegistrationResponseEntity
-import com.example.domain.entity.RegistrationResultEntity
-import com.example.domain.entity.ResetCompleteDataEntity
-import com.example.domain.entity.ResetCompleteErrorEntity
-import com.example.domain.entity.ResetCompleteErrorsEntity
-import com.example.domain.entity.ResetCompleteResponseEntity
-import com.example.domain.entity.ResetCompleteResultEntity
-import com.example.domain.entity.SendCodeDataErrorEntity
-import com.example.domain.entity.SendCodeErrorEntity
-import com.example.domain.entity.SendCodeErrorsEntity
-import com.example.domain.entity.SendCodeResponseEntity
-import com.example.domain.entity.SendCodeResultEntity
+import com.example.domain.entity.responses.DataCompleteResponseEntity
+import com.example.domain.entity.responses.DataEmailResetEntity
+import com.example.domain.entity.responses.DataSendCodeDomain
+import com.example.domain.entity.responses.EmailPassDataErrorEntity
+import com.example.domain.entity.responses.EmailPassResetErrorEntity
+import com.example.domain.entity.responses.EmailPassResetErrorsEntity
+import com.example.domain.entity.responses.EmailResetResponseEntity
+import com.example.domain.entity.results.EmailResetResultEntity
+import com.example.domain.entity.responses.ErrorResponse
+import com.example.domain.entity.responses.LoginData
+import com.example.domain.entity.responses.LoginDataError
+import com.example.domain.entity.responses.LoginErrorsDomain
+import com.example.domain.entity.responses.LoginResponse
+import com.example.domain.entity.results.LoginResultEntity
+import com.example.domain.entity.responses.LoginTokens
+import com.example.domain.entity.responses.RegistrationDataEntity
+import com.example.domain.entity.responses.RegistrationErrorDetailEntity
+import com.example.domain.entity.responses.RegistrationErrorEntity
+import com.example.domain.entity.responses.RegistrationErrorsDataEntity
+import com.example.domain.entity.responses.RegistrationResponseEntity
+import com.example.domain.entity.results.RegistrationResultEntity
+import com.example.domain.entity.responses.ResetCompleteDataEntity
+import com.example.domain.entity.responses.ResetCompleteErrorEntity
+import com.example.domain.entity.responses.ResetCompleteErrorsEntity
+import com.example.domain.entity.responses.ResetCompleteResponseEntity
+import com.example.domain.entity.results.ResetCompleteResultEntity
+import com.example.domain.entity.responses.SendCodeDataErrorEntity
+import com.example.domain.entity.responses.SendCodeErrorEntity
+import com.example.domain.entity.responses.SendCodeErrorsEntity
+import com.example.domain.entity.responses.SendCodeResponseEntity
+import com.example.domain.entity.results.SendCodeResultEntity
 import com.example.domain.repository.AppRepository
 import com.squareup.moshi.Moshi
 import kotlinx.coroutines.flow.firstOrNull
