--- domain/src/main/java/com/example/domain/entity/DataCompleteResponseEntity.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class DataCompleteResponseEntity(
-    val success: String,
-)
