--- domain/src/main/java/com/example/domain/entity/DataEmailResetEntity.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class DataEmailResetEntity(
-    val success: String
-)
