--- domain/src/main/java/com/example/domain/entity/DataSendCodeDomain.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class DataSendCodeDomain(
-    val success: String
-)
