--- domain/src/main/java/com/example/domain/entity/EmailPassDataErrorEntity.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-data class EmailPassDataErrorEntity(
-    val errors: List<EmailPassResetErrorsEntity>,
-    val type: String
-)
