--- domain/src/main/java/com/example/domain/entity/EmailPassResetErrorEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class EmailPassResetErrorEntity(
-    val code: Int,
-    val `data`: EmailPassDataErrorEntity,
-    val message: Any? = null,
-    val status: String
-)
