--- domain/src/main/java/com/example/domain/entity/EmailPassResetErrorsEntity.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class EmailPassResetErrorsEntity(
-    val detail: String
-)
