--- domain/src/main/java/com/example/domain/entity/EmailResetResponseEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class EmailResetResponseEntity(
-    val code: Int,
-    val `data`: DataEmailResetEntity,
-    val message: Any? = null,
-    val status: String
-)
