--- domain/src/main/java/com/example/domain/entity/ErrorResponse.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class ErrorResponse(
-    val code: Int,
-    val `data`: LoginDataError,
-    val message: Any? = null,
-    val status: String
-)
