--- domain/src/main/java/com/example/domain/entity/LoginData.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-    data class LoginData(
-        val email: String,
-        val tokens: LoginTokens,
-    )
