--- domain/src/main/java/com/example/domain/entity/LoginDataError.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-data class LoginDataError(
-    val errors: List<LoginErrorsDomain>,
-    val type: String
-)
