--- domain/src/main/java/com/example/domain/entity/LoginErrorsDomain.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class LoginErrorsDomain(
-    val detail: String
-)
