--- domain/src/main/java/com/example/domain/entity/LoginResponse.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class LoginResponse(
-    val code: Int,
-    val `data`: LoginData,
-    val message: Any? = null,
-    val status: String
-)
