--- domain/src/main/java/com/example/domain/entity/LoginTokens.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-data class LoginTokens(
-    val access: String,
-    val refresh: String
-)
