--- domain/src/main/java/com/example/domain/entity/ResetCompleteDataEntity.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-data class ResetCompleteDataEntity(
-    val errors: List<ResetCompleteErrorsEntity>,
-    val type: String
-)
