--- domain/src/main/java/com/example/domain/entity/ResetCompleteErrorEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class ResetCompleteErrorEntity(
-    val code: Int,
-    val `data`: ResetCompleteDataEntity,
-    val message: Any? = null,
-    val status: String
-)
