--- domain/src/main/java/com/example/domain/entity/ResetCompleteErrorsEntity.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class ResetCompleteErrorsEntity(
-    val detail: String
-)
