--- domain/src/main/java/com/example/domain/entity/ResetCompleteResponseEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class ResetCompleteResponseEntity(
-    val code: Int,
-    val `data`: DataCompleteResponseEntity,
-    val message: Any? = null,
-    val status: String,
-)
