--- domain/src/main/java/com/example/domain/entity/SendCodeDataErrorEntity.kt
@@ -1,6 +1,0 @@
-package com.example.domain.entity
-
-data class SendCodeDataErrorEntity(
-    val errors: List<SendCodeErrorsEntity>,
-    val type: String
-)
