--- domain/src/main/java/com/example/domain/entity/SendCodeErrorEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class SendCodeErrorEntity(
-    val code: Int,
-    val `data`: SendCodeDataErrorEntity,
-    val message: Any? = null,
-    val status: String
-)
