--- domain/src/main/java/com/example/domain/entity/SendCodeErrorsEntity.kt
@@ -1,5 +1,0 @@
-package com.example.domain.entity
-
-data class SendCodeErrorsEntity(
-    val detail: String
-)
