--- domain/src/main/java/com/example/domain/entity/SendCodeResponseEntity.kt
@@ -1,8 +1,0 @@
-package com.example.domain.entity
-
-data class SendCodeResponseEntity(
-    val code: Int,
-    val `data`: DataSendCodeDomain,
-    val message: Any? = null,
-    val status: String
-)
