+++ domain/src/main/java/com/example/domain/entity/responses/SendCodeResponseEntity.kt
@@ -1,0 +1,28 @@
+package com.example.domain.entity.responses
+
+data class SendCodeResponseEntity(
+    val code: Int,
+    val `data`: DataSendCodeDomain,
+    val message: Any? = null,
+    val status: String
+)
+
+data class SendCodeErrorsEntity(
+    val detail: String
+)
+
+data class SendCodeErrorEntity(
+    val code: Int,
+    val `data`: SendCodeDataErrorEntity,
+    val message: Any? = null,
+    val status: String
+)
+
+data class SendCodeDataErrorEntity(
+    val errors: List<SendCodeErrorsEntity>,
+    val type: String
+)
+
+data class DataSendCodeDomain(
+    val success: String
+)
