+++ domain/src/main/java/com/example/domain/entity/results/EmailResetResultEntity.kt
@@ -1,4 +1,7 @@
-package com.example.domain.entity
+package com.example.domain.entity.results
+
+import com.example.domain.entity.responses.DataEmailResetEntity
+import com.example.domain.entity.responses.EmailPassResetErrorsEntity
 
 sealed class EmailResetResultEntity {
     data class Success(val data: DataEmailResetEntity) : EmailResetResultEntity()
