+++ domain/src/main/java/com/example/domain/entity/results/LoginResultEntity.kt
@@ -1,4 +1,7 @@
-package com.example.domain.entity
+package com.example.domain.entity.results
+
+import com.example.domain.entity.responses.LoginData
+import com.example.domain.entity.responses.LoginErrorsDomain
 
 sealed class LoginResultEntity {
     data class Success(val data: LoginData) : LoginResultEntity()
