+++ domain/src/main/java/com/example/domain/entity/results/RegistrationResultEntity.kt
@@ -1,5 +1,7 @@
-package com.example.domain.entity
+package com.example.domain.entity.results
 
+import com.example.domain.entity.responses.RegistrationDataEntity
+import com.example.domain.entity.responses.RegistrationErrorDetailEntity
 
 sealed class RegistrationResultEntity {
  data class Success(val data: RegistrationDataEntity) : RegistrationResultEntity()
