+++ domain/src/main/java/com/example/domain/entity/results/ResetCompleteResultEntity.kt
@@ -1,4 +1,7 @@
-package com.example.domain.entity
+package com.example.domain.entity.results
+
+import com.example.domain.entity.responses.ResetCompleteErrorsEntity
+import com.example.domain.entity.responses.DataCompleteResponseEntity
 
 sealed class ResetCompleteResultEntity {
     data class Success(val data: DataCompleteResponseEntity) : ResetCompleteResultEntity()
