+++ domain/src/main/java/com/example/domain/entity/results/SendCodeResultEntity.kt
@@ -1,4 +1,7 @@
-package com.example.domain.entity
+package com.example.domain.entity.results
+
+import com.example.domain.entity.responses.DataSendCodeDomain
+import com.example.domain.entity.responses.SendCodeErrorsEntity
 
 sealed class SendCodeResultEntity {
     data class Success(val data: DataSendCodeDomain) : SendCodeResultEntity()
